requirements.txt
# ===============================
# Core Data & ML
# ===============================
pandas>=2.2.2
numpy>=1.26.2
scikit-learn>=1.3.2
joblib>=1.3.2

# ===============================
# Visualization (optional)
# ===============================
matplotlib>=3.8.1
seaborn>=0.12.2

# ===============================
# API & Serving
# ===============================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.1

# ===============================
# MLOps / Experiment Tracking
# ===============================
mlflow>=2.9.2
pyarrow>=14.0.2
python-dotenv>=1.0.0

# ===============================
# Utils
# ===============================
pyyaml>=6.0
tqdm>=4.66.1
python-dateutil>=2.8.2
user-agents>=2.2.0
jinja2>=3.1.2

# ===============================
# Reports (optional)
# ===============================
pdfkit>=1.0.0

